requirements.txt
anyio==3.6.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.78.0
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
pydantic==1.9.1
python-dotenv==0.20.0
requests==2.28.1
requests-async==0.6.2
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
